Changed css files to sass files. And restructured the css to use nested styles. should be used like this to improve readability
