+++ src/Assets/styles/variables.scss
@@ -1,0 +1,16 @@
+// Global Color Palette
+$background: #ebebeb;
+$white: #ffffff;
+$slightyWhite: #fafafa;
+$offWhite: #dcdde1;
+$lightGray: #9a9a9a;
+$gray: #858d90;
+$grayDarker: #353e43;
+$orange: #FF7F00;
+$blue: #47B5FF;
+$navyBlue: #1e4364;
+$darkerNavyBlue: #17334f;
+$lightGreen: #2ecc71;
+$green: #27ae60;
+$darkBlue: #06283D;
+$black: #000;
