+++ src/Components/Button/Button.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import "./button.css";
+import "./button.scss";
 
 const Button = ({ label, onClick, onMouseEnter, onMouseLeave }) => {
   return (
