+++ src/Components/Button/button.scss
@@ -1,15 +1,17 @@
+@use '../../Assets/styles/variables.scss' as v;
+
 .navBtn {
     
     background-color: unset;
     border: none;
     outline: none;
-    color: white;
+    color: v.$white;
     cursor: pointer;
     font-size: 16px;
     padding: 16px;
 }
 .navBtn:hover{
-    background: #3c6382;
+    background: v.$navyBlue;
     transition: .4s;
     border-radius: 4px;
 }
