--- src/Components/Herobanner/herobanner.css
@@ -1,54 +1,0 @@
-.heroWrapper {
-    display: flex;
-    /* background-image: linear-gradient(to bottom,#06283D, #1363DF,#47B5FF,#dcdde1 98%); */
-    align-items: center;
-    justify-content: center;
-    max-height: 500px;
-    height: 100%;
-    padding: 101px 0 100px 0;
-    flex-direction: column;
-  }
-  
-  .heroWrapper a {
-    max-width: 410px;
-    max-height: 300px;
-    width: 100%;
-    height: 100%;
-  }
-
-  .heroWrapper img {
-
-    width: 100%;
-    height: 100%;
-  }
-
-  .heroWrapper h1 {
-    color: white;
-  }
-  
-  .first-letter {
-    color: #FF7F00;
-  }
-
-  .heroWrapper h2 {
-    color: white;
-  }
-
-  .h2-wrapper {
-    display: flex;
-    align-items: center;
-  }
-  
-  .h2-wrapper h2 {
-    margin: 0;
-    font: 10px;
-  }
-
-  .h2-first {
-    color: #FF7F00 !important;
-  }
-  
-  .h2-second {
-    font-size: large;
-    padding-top: 5px;
-  }
