+++ src/Components/Herobanner/herobanner.scss
@@ -1,0 +1,55 @@
+@use '../../Assets/styles/variables.scss' as v;
+
+.heroWrapper {
+    display: flex;
+    /* background-image: linear-gradient(to bottom,#06283D, #1363DF,#47B5FF,#dcdde1 98%); */
+    align-items: center;
+    justify-content: center;
+    max-height: 500px;
+    height: 100%;
+    padding: 101px 0 100px 0;
+    flex-direction: column;
+    
+    a {
+      max-width: 410px;
+      max-height: 300px;
+      width: 100%;
+      height: 100%;
+    }
+    
+    img {
+      width: 100%;
+      height: 100%;
+    }
+    
+    h1, h2 {
+      color: v.$white;
+    }
+    
+    .first-letter {
+      color: v.$orange;
+    }
+    
+    .h2-wrapper {
+      display: flex;
+      align-items: center;
+    }
+    
+    .h2-wrapper h2 {
+      margin: 0;
+      font: 10px;
+    }
+  
+    .h2-first {
+      color: v.$orange !important;
+    }
+    
+    .h2-second {
+      font-size: large;
+      padding-top: 5px;
+    }
+  }
+  
+
+
+
