--- src/Components/Navbar/navbar.css
@@ -1,85 +1,0 @@
-.navbar {
-  list-style: none;
-  margin: 0;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end; 
-  top: 0;
-  right:0; 
-  padding: 0;
-  position: absolute;
-
-  width: 40%;
-  background-color: #17334f;
-  margin-top: 16px;
-  justify-content: center;
-}
-
-ul.navbar:before {
-  content: "";
-  position: absolute;
-  top: 0;
-  right: auto;
-  bottom: auto;
-  left: -12px;
-  height: 100%;
-  width: 25px;
-  background: #17334f;
-  -webkit-transform: skew(-20deg,0);
-  -moz-transform: skew(-20deg,0);
-  -ms-transform: skew(-20deg,0);
-  transform: skew(-20deg,0);
-}
-
-.navbar li {
-  list-style: none;
-}
-
-.navbar a {
-  display: block;
-  font-size: medium;
-  color: white;
-  text-decoration: none;
-  font-family: Montserrat, Arial, sans-serif;
-  font-weight: bold;
-  padding: 10px;
-}
-
-.navbar-submenu {
-  display: none;
-  position: absolute;
-  background-color: #173350;
-  padding: 12px;
-  transition: 0.4;
-}
-
-.navbar-submenu a {
-  padding: 0;
-}
-
-.navbar-button-wrapper:hover .navbar-submenu {
-  display: block;
-}
-
-
-.navbar-submenu2 {
-  display: none;
-  position: absolute;
-  background-color: #173350;
-  padding: 12px;
-  transition: 0.4;
-}
-
-.navbar-submenu2 a {
-  padding: 0;
-}
-
-.navbar-button-wrapper:hover .navbar-submenu2 {
-  display: block;
-}
-
-.instagram-logo {
-  width: 34px;
-  height: 34px;
-}
-
