+++ src/Components/Navbar/navbar.scss
@@ -1,0 +1,73 @@
+@use '../../Assets/styles/variables.scss' as v;
+
+.navbar {
+  list-style: none;
+  margin: 0;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end; 
+  top: 0;
+  right:0; 
+  padding: 0;
+  position: absolute;
+
+  width: 40%;
+  background-color: v.$darkerNavyBlue;
+  margin-top: 16px;
+  justify-content: center;
+
+  
+  .instagram-logo {
+    width: 34px;
+    height: 34px;
+  }
+
+  .navbar-item {
+    list-style: none;
+
+    a {
+      display: block;
+      font-size: medium;
+      color: v.$white;
+      text-decoration: none;
+      font-family: Montserrat, Arial, sans-serif;
+      font-weight: bold;
+      padding: 10px;
+    }
+    
+    .navbar-button-wrapper {
+      
+      .navbar-submenu {
+        display: none;
+        position: absolute;
+        background-color: v.$darkerNavyBlue;
+        padding: 12px;
+        transition: 0.4;
+        
+        a {
+          padding: 0;
+        }
+      }
+
+      &:hover .navbar-submenu {
+        display: block;
+      }
+    }
+  }
+}
+
+.navbar:before {
+  content: "";
+  position: absolute;
+  top: 0;
+  right: auto;
+  bottom: auto;
+  left: -12px;
+  height: 100%;
+  width: 25px;
+  background: v.$darkerNavyBlue;
+  -webkit-transform: skew(-20deg,0);
+  -moz-transform: skew(-20deg,0);
+  -ms-transform: skew(-20deg,0);
+  transform: skew(-20deg,0);
+}
