+++ src/Components/Programpackages/PackagesSection.jsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from 'react';
 import sanityClient from '../../client.js'
-import './packagessection.css';
+import './packagessection.scss';
 import PackageItem from './PackageItem';
 
 const PackagesSection = () => {
