--- src/Components/Programpackages/packagessection.css
@@ -1,132 +1,0 @@
-
-
-h1 {
-    color: #9a9a9a;
-    text-align: center;
-    text-transform: uppercase;
-    font-size: 44px;
-}
-
-.packs-wrapper {
-    margin-top: 800px;
-    padding: 32px;
-    padding-right: 32px;
-    background-image: linear-gradient(to bottom, #16314D, #1E4364, #28547B, #336791);
-    margin: 200px auto 0px auto;
-    
-}
-
-.card-list {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    column-gap: 20px;
-    margin: 0 auto;
-    list-style: none;
-    padding: 0;
-}
-
-.card-list-item:nth-child(2) {
-    margin-top: -8px;
-    margin-bottom: 8px;
-}
-
-.card-list-item {
-    background-color: #f5f5f5;
-    min-width: 200px;
-    display: flex;
-    flex-direction: column;
-    
-}
-
-.headline-wrapper, .price-wrapper {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-}
-
-.headline-wrapper {
-    background-color: white;
-}
-
-.headline-wrapper h2 {
-    font-size: 32px;
-    text-transform: uppercase;
-    margin: 0;
-    padding-top: 16px;
-    color: #06283D;
-}
-
-.headline-wrapper {
-    margin: 0;
-}
-
-.price-wrapper {
-    padding-top: 16px;
-}
-
-.price-wrapper h3 {
-    font-size: 40px;
-    text-transform: uppercase;
-}
-
-.price-wrapper h3, .price-wrapper p  {
-    margin: 0;
-}
-
-.info-list {
-    list-style: none;
-    padding: 16px;
-}
-
-.info-item {
-    border-bottom: 2px solid #c3c3c3;
-    padding: 0;
-    display: flex;
-    justify-content: center;
-    flex-direction: row;
-    align-items: center;
-    gap: 8px;
-}
-
-.dot {
-    max-width: 16px;
-    max-height: 16px;
-    width: 16px;
-    height: 16px;
-    background-color: #7b7b7b;
-    border-radius: 50%;
-}
-
-.dot img {
-    height: 100%;
-    width: 100%;
-}
-
-.info-item p {
-   margin: 16px 0;
-}
-
-.card-wrapper {
-    margin: auto;
-    max-width: 1200px;
-}
-
-.button-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: end;
-    padding-bottom: 1rem;
-    height: 100%;
-}
-
-.button-wrapper button {
-    background-color: #27ae60;
-    padding: 1rem 3rem;
-    border-radius: 4px;
-}
-
-.button-wrapper button:hover {
-    transition: 0.5s;
-    background-color: #2ecc71;
-}
