+++ src/Components/Programpackages/packagessection.scss
@@ -1,0 +1,123 @@
+@use '../../Assets/styles/variables.scss' as v;
+
+h1 {
+    color: v.$lightGray;
+    text-align: center;
+    text-transform: uppercase;
+    font-size: 44px;
+}
+
+.packs-wrapper {
+    margin-top: 800px;
+    padding: 32px;
+    padding-right: 32px;
+    background-image: linear-gradient(to bottom, v.$darkerNavyBlue, v.$navyBlue);
+    margin: 200px auto 0px auto;
+    
+    .card-list {
+        display: grid;
+        grid-template-columns: 1fr 1fr 1fr;
+        column-gap: 20px;
+        margin: 0 auto;
+        list-style: none;
+        padding: 0;
+        
+        .card-list-item {
+            background-color: v.$slightyWhite;
+            min-width: 200px;
+            display: flex;
+            flex-direction: column;
+            
+            &:nth-child(2) {
+                margin-top: -8px;
+                margin-bottom: 8px;
+            }
+            
+            .headline-wrapper, .price-wrapper {
+                display: flex;
+                align-items: center;
+                justify-content: center;
+                flex-direction: column;
+            }
+            .headline-wrapper {
+                background-color: v.$white;
+                margin: 0;
+
+                h2 {
+                    font-size: 32px;
+                    text-transform: uppercase;
+                    margin: 0;
+                    padding-top: 16px;
+                    color: #06283D;
+                }
+            }
+
+            .price-wrapper {
+                padding-top: 16px;
+
+                h3 {
+                    font-size: 40px;
+                    text-transform: uppercase;
+                }
+
+                h3, p {
+                    margin: 0;
+                }
+            }    
+
+            .info-list {
+                list-style: none;
+                padding: 16px;
+                
+                .info-item {
+                    border-bottom: 2px solid v.$offWhite;
+                    padding: 0;
+                    display: flex;
+                    justify-content: center;
+                    flex-direction: row;
+                    align-items: center;
+                    gap: 8px;
+
+                    .dot {
+                        max-width: 16px;
+                        max-height: 16px;
+                        width: 16px;
+                        height: 16px;
+                        background-color: v.$gray;
+                        border-radius: 50%;
+                        
+                        img {
+                            height: 100%;
+                            width: 100%;
+                        }
+
+                        p {
+                            margin: 16px 0;
+                        }
+                    }
+                }
+            }
+            .button-wrapper {
+                display: flex;
+                justify-content: center;
+                align-items: flex-end;
+                padding-bottom: 1rem;
+                height: 100%;
+                
+                button {
+                    background-color: v.$green;
+                    padding: 1rem 3rem;
+                    border-radius: 4px;
+                
+                    &:hover {
+                        transition: 0.5s;
+                        background-color: v.$lightGreen;
+                    }
+                }
+            }
+        }
+    }
+}
+
+
+
