+++ src/Pages/Home.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Navbar } from '../Components/Navbar/Navbar';
-import './Home.css';
 import { HeroBanner } from '../Components/Herobanner/HeroBanner';
 // import ImageWithText from '../Components/ImageWithText/ImageWithText'; //Can stay, may use later.
 import PackagesSection from '../Components/Programpackages/PackagesSection';
