+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import {BrowserRouter} from "react-router-dom"
-import "./index.css"
+import "./index.scss"
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
